Model changes (#22)
* Refactor inventory (#21)

* [ci skip]  Add link to published website from repo

Added a text under title

* Refactor inventory

* Add instructions to annotate SystemData class

* Fix SystemData indentation

* mmodel changes
