+++ README.adoc
@@ -162,7 +162,8 @@ served are now more meaningful:
       example: foo
     responses:
       404:
-        description: Missing description - to be filtered.
+        description: Invalid hostname or the system service may not be running on
+          the particular host.
         content:
           text/plain: {}
       200:
@@ -188,12 +189,20 @@ served are now more meaningful:
 
 OpenAPI annotations can also be added to POJOs to describe what they represent. Currently, your OpenAPI 
 document doesn't have a very meaningful description of the `InventoryList` POJO and hence it's very 
-difficult to tell excatly what that POJO is used for. To describe the `InventoryList` POJO in more detail, augment the 
+difficult to tell exactly what that POJO is used for. To describe the `InventoryList` POJO in more detail, augment the 
 `src/main/java/io/openliberty/guides/inventory/model/InventoryList.java` file with some OpenAPI annotations:
 
 [source, Java]
 ----
 include::finish/src/main/java/io/openliberty/guides/inventory/model/InventoryList.java[tags=**;!copyright]
+----
+
+Likewise, annotate the `src/main/java/io/openliberty/guides/inventory/model/SystemData.java` POJO, 
+which is referenced in the `InventoryList` class:
+
+[source, Java]
+----
+include::finish/src/main/java/io/openliberty/guides/inventory/model/SystemData.java[tags=**;!copyright]
 ----
 
 Again, run the `mvn compile` command and refresh the {openapi-url} URL to see the updated OpenAPI tree:
@@ -210,15 +219,11 @@ components:
         systems:
           type: array
           items:
-            $ref: '#/components/schemas/System'
+            $ref: '#/components/schemas/SystemData'
         total:
           type: integer
       description: POJO that represents the inventory contents.
-    Properties:
-      type: object
-      additionalProperties:
-        type: string
-    System:
+    SystemData:
       required:
       - hostname
       - properties
@@ -231,6 +236,10 @@ components:
           additionalProperties:
             type: string
       description: POJO that represents a single inventory entry.
+    Properties:
+      type: object
+      additionalProperties:
+        type: string
 ----
 
 
@@ -285,12 +294,13 @@ info:
   license:
     name: Eclipse Public License - v 1.0
     url: https://www.eclipse.org/legal/epl-v10.html
-  version: "1.0"
+  version: 1.0.0
 servers:
 - url: http://localhost:{port}
   description: Simple Open Liberty.
   variables:
     port:
+      default: "9080"
       description: Server HTTP port.
 ----
 
@@ -339,8 +349,6 @@ Run the `mvn compile` command and refresh the {openapi-url} URL to see the updat
 ----
 /inventory/properties:
   get:
-    summary: Get JVM system properties.
-    description: Returns the JVM system properties for the host running this service.
     operationId: getProperties
     responses:
       200:
@@ -348,7 +356,9 @@ Run the `mvn compile` command and refresh the {openapi-url} URL to see the updat
         content:
           application/json:
             schema:
-              $ref: '#/components/schemas/Properties'
+              type: object
+              additionalProperties:
+                type: string
 ----
