+++ start/src/main/java/io/openliberty/guides/inventory/model/SystemData.java
@@ -12,25 +12,33 @@
 // end::copyright[]
 package io.openliberty.guides.inventory.model;
 
-import java.util.List;
-
+import java.util.Properties;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
 
-@Schema(name="InventoryList", description="POJO that represents the inventory contents.")
-public class InventoryList {
+public class SystemData {
 
-    @Schema(required = true)
-    private List<SystemData> systems;
+    private final String hostname;
 
-    public InventoryList(List<SystemData> systems) {
-        this.systems = systems;
+    private final Properties properties;
+
+    public SystemData(String hostname, Properties properties) {
+        this.hostname = hostname;
+        this.properties = properties;
     }
 
-    public List<SystemData> getSystems() {
-        return systems;
+    public String getHostname() {
+        return hostname;
     }
 
-    public int getTotal() {
-        return systems.size();
+    public Properties getProperties() {
+        return properties;
+    }
+
+    @Override
+    public boolean equals(Object host) {
+        if (host instanceof SystemData) {
+            return hostname.equals(((SystemData) host).getHostname());
+        }
+        return false;
     }
 }
